almost syncs

+++ packages/sync-core/src/db/sql/schema/create_sync_tables.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS action_records (
 	id TEXT PRIMARY KEY DEFAULT gen_random_uuid(),
 	_tag TEXT NOT NULL,
 	client_id TEXT NOT NULL,
-	transaction_id BIGINT NOT NULL,
+	transaction_id FLOAT NOT NULL,
 	clock JSONB NOT NULL,
 	args JSONB NOT NULL,
 	created_at TIMESTAMP WITH TIME ZONE DEFAULT NOW(),
